requirements.txt
Django==1.5.4
django-contact-form==1.0
django-ckeditor==4.0.2
django-forms-builder==0.8.5
django-preferences==0.0.6
django-toolbelt==0.0.1
django-widget-tweaks==1.3
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
mimeparse
mock
psycopg2==2.5.1
pytz
South==0.8.1
static==0.4
wsgiref==0.1.2
